Restructure from pantry to stack.lock
