+++ subs/pantry/src/Pantry/Types.hs
@@ -41,7 +41,7 @@ module Pantry.Types
   , renderTree
   , parseTree
   , SHA256
-  , Unresolved
+  , Unresolved (..)
   , resolvePaths
   , Package (..)
   , PackageCabal (..)
@@ -58,6 +58,7 @@ module Pantry.Types
   , toRawArchive
   , Repo (..)
   , RepoType (..)
+  , GitHubRepo (..)
   , parsePackageIdentifier
   , parsePackageName
   , parsePackageNameThrowing
@@ -66,17 +67,13 @@ module Pantry.Types
   , parseVersionThrowing
   , packageIdentifierString
   , packageNameString
-  , resolveLockFile
-  , resolveSnapshotFile
-  , resolveSnapshotLockFile
-  , LockFile (..)
-  , loadLockFile
-  , loadSnapshotLockFile
   , parseAndResolvePackageLocation
+  , parseRawSnapshotLocationPath
   , flagNameString
   , versionString
   , moduleNameString
   , OptionalSubdirs (..)
+  , osToRpms
   , ArchiveLocation (..)
   , RelFilePath (..)
   , CabalString (..)
@@ -97,6 +94,8 @@ module Pantry.Types
   , SnapshotLocation (..)
   , toRawSL
   , parseRawSnapshotLocation
+  , parseArchiveLocationText
+  , parseArchiveLocationObject
   , RawSnapshotLayer (..)
   , SnapshotLayer (..)
   , toRawSnapshotLayer
@@ -106,6 +105,7 @@ module Pantry.Types
   , SnapshotPackage (..)
   , parseWantedCompiler
   , RawPackageMetadata (..)
+  , rpmEmpty
   , PackageMetadata (..)
   , toRawPM
   , cabalFileName
@@ -1538,128 +1538,6 @@ parseAndResolvePackageLocation rootDir v = do
   (Unresolved unresolvedPL) <- parsePackageLocation v
   pure $ unresolvedPL (Just rootDir)
 
-combineUnresolved :: Unresolved a -> Unresolved b -> Unresolved (a,b)
-combineUnresolved a b = do
-  ua <- a
-  ub <- b
-  pure (ua, ub)
-
-parseSingleObject :: Value -> Parser (Unresolved (PackageLocation, RawPackageLocation))
-parseSingleObject value = withObject "LockFile" (\obj -> do
-                                                   original <- obj .: "original"
-                                                   complete <- obj .: "complete"
-                                                   orig <- parseRPL original
-                                                   comp <- parsePImmutable complete
-                                                   pure $ combineUnresolved comp orig
-                                                   ) value
-
-data LockFile a = LockFile {
-      lfPackageLocation :: a [(PackageLocation, RawPackageLocation)],
-      lfoResolver :: a RawSnapshotLocation,
-      lfcResolver :: a SnapshotLocation
-}
-
-parseImmutableObject :: Value -> Parser (Unresolved (PackageLocationImmutable, RawPackageLocationImmutable))
-parseImmutableObject value = withObject "LockFile" (\obj -> do
-                                                   original <- obj .: "original"
-                                                   complete <- obj .: "complete"
-                                                   orig <- parseRPLI original
-                                                   comp <- parsePLI complete
-                                                   pure $ combineUnresolved comp orig
-                                                   ) value
-
-
-parseSnapshotLockFile ::
-       Value -> Parser (Unresolved [(PackageLocationImmutable, RawPackageLocationImmutable)])
-parseSnapshotLockFile =
-    withObject
-        "SnapshotLockFile"
-        (\obj -> do
-             vals <- obj .: "dependencies"
-             xs <-
-                 withArray
-                     "SnapshotLockArray"
-                     (\vec -> sequence $ Vector.map parseImmutableObject vec)
-                     vals
-             pure $ sequence $ Vector.toList xs)
-
-resolveSnapshotLockFile :: Path Abs Dir -> Value -> Parser (IO [(PackageLocationImmutable, RawPackageLocationImmutable)])
-resolveSnapshotLockFile rootDir val = do
-  pkgs <- parseSnapshotLockFile val
-  let pkgsLoc = resolvePaths (Just rootDir) pkgs
-  pure pkgsLoc
-
-loadSnapshotLockFile :: Path Abs File -> Path Abs Dir -> IO [(PackageLocationImmutable, RawPackageLocationImmutable)]
-loadSnapshotLockFile lockFile rootDir = do
-  val <- Yaml.decodeFileThrow (toFilePath lockFile)
-  case Yaml.parseEither (resolveSnapshotLockFile rootDir) val of
-    Left str -> fail $ "Cannot parse snapshot lock file: Got error " <> str <> (show val)
-    Right lockFileIO -> lockFileIO
-
-parseSnapshotFile :: Value -> Parser (Unresolved ([RawPackageLocationImmutable], RawSnapshotLocation))
-parseSnapshotFile (Object obj) = do
-  packages <- obj .: "packages"
-  resolver <- obj .: "resolver"
-  xs <- withArray "SnapshotFileArray" (\vec -> sequence $ Vector.map parseRPLI vec) packages
-  resolver <- parseRSL resolver
-  pure $ combineUnresolved (sequence $ Vector.toList xs) resolver
-parseSnapshotFile val = fail $ "Expected Object, but got: " <> (show val)
-
-resolveSnapshotFile :: Path Abs Dir -> Value -> Parser (IO ([RawPackageLocationImmutable], RawSnapshotLocation))
-resolveSnapshotFile rootDir val = do
-  unrpl <- parseSnapshotFile val
-  let pkgLoc = resolvePaths (Just rootDir) unrpl
-  pure pkgLoc
-
-parseLockFile ::
-       Value -> Parser (LockFile Unresolved)
-parseLockFile value = withObject "LockFile" (\obj -> do
-                                               vals :: Value <- obj .: "dependencies"
-                                               xs <- withArray "LockFileArray" (\vec -> sequence $ Vector.map parseSingleObject vec) vals
-                                               resolver <- obj .: "resolver"
-                                               roriginal <- resolver .: "original"
-                                               rcomplete <- resolver .: "complete"
-                                               ro <- parseRSL roriginal
-                                               rc <- parseSL rcomplete
-                                               pure $ LockFile {
-                                                          lfPackageLocation = sequence (Vector.toList xs),
-                                                          lfoResolver = ro,
-                                                          lfcResolver = rc
-                                                        }
-                                            ) value
-
-resolveLockFile :: Path Abs Dir -> Value -> Parser (LockFile IO)
-resolveLockFile rootDir v = do
-    lockFile <- parseLockFile v
-    let pkgLoc = (lfPackageLocation lockFile)
-        origRes = lfoResolver lockFile
-        compRes = lfcResolver lockFile
-        pkgLoc' = resolvePaths (Just rootDir) pkgLoc
-    pure $
-        LockFile
-            { lfPackageLocation = pkgLoc'
-            , lfoResolver = resolvePaths (Just rootDir) origRes
-            , lfcResolver = resolvePaths (Just rootDir) compRes
-            }
-
-loadLockFile :: Path Abs File -> IO (LockFile IO)
-loadLockFile lockFile = do
-  val <- Yaml.decodeFileThrow (toFilePath lockFile)
-  case Yaml.parseEither (resolveLockFile (parent lockFile)) val of
-    Left str -> fail $ "Cannot parse lock file: Got error " <> str
-    Right lockFileIO -> pure lockFileIO
-
-
-parsePImmutable :: Value -> Parser (Unresolved PackageLocation)
-parsePImmutable v = do
-  xs :: Unresolved PackageLocationImmutable <- parseJSON v
-  pure $ PLImmutable <$> xs
-
-parsePLI :: Value -> Parser (Unresolved PackageLocationImmutable)
-parsePLI v = do
-  x <- parseJSON v
-  pure x
-
 instance FromJSON (Unresolved PackageLocationImmutable) where
     parseJSON v = repoObject v <|> archiveObject v <|> hackageObject v <|> github v
                   <|> fail ("Could not parse a UnresolvedPackageLocationImmutable from: " ++ show v)
@@ -1710,106 +1588,6 @@ instance FromJSON (Unresolved PackageLocationImmutable) where
               archiveSize <- o .: "size"
               archiveSubdir <- o .: "subdir"
               pure $ pure $ PLIArchive Archive {..} pm) value
-
-parseRPLImmutable :: Value -> Parser (Unresolved RawPackageLocation)
-parseRPLImmutable v = do
-  xs :: Unresolved RawPackageLocationImmutable <- parseRPLI v
-  pure $ RPLImmutable <$> xs
-
-parseRPL :: Value -> Parser (Unresolved RawPackageLocation)
-parseRPL v = parseRPLImmutable v <|> parseResolvedPath v
-
-parseRPLI :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseRPLI v =
-    parseRPLHttpText v <|> parseRPLHackageText v <|> parseRPLHackageObject v <|>
-    parseRPLRepo v <|>
-    parseArchiveRPLObject v <|>
-    parseGithubRPLObject v
-
-parseResolvedPath :: Value -> Parser (Unresolved RawPackageLocation)
-parseResolvedPath value = mkMutable <$> parseJSON value
-    where
-      mkMutable :: Text -> Unresolved RawPackageLocation
-      mkMutable t = Unresolved $ \mdir -> do
-        case mdir of
-          Nothing -> throwIO $ MutablePackageLocationFromUrl t
-          Just dir -> do
-            abs' <- resolveDir dir $ T.unpack t
-            pure $ RPLMutable $ ResolvedPath (RelFilePath t) abs'
-
-parseRPLHttpText :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseRPLHttpText = withText "UnresolvedPackageLocationImmutable.UPLIArchive (Text)" $ \t ->
-        case parseArchiveLocationText t of
-          Nothing -> fail $ "Invalid archive location: " ++ T.unpack t
-          Just (Unresolved mkArchiveLocation) ->
-            pure $ Unresolved $ \mdir -> do
-              raLocation <- mkArchiveLocation mdir
-              let raHash = Nothing
-                  raSize = Nothing
-                  raSubdir = T.empty
-              pure $ RPLIArchive RawArchive {..} rpmEmpty
-
-parseRPLHackageText :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseRPLHackageText = withText "UnresolvedPackageLocationImmutable.UPLIHackage (Text)" $ \t ->
-        case parsePackageIdentifierRevision t of
-          Left e -> fail $ show e
-          Right pir -> pure $ pure $ RPLIHackage pir Nothing
-
-parseRPLHackageObject :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseRPLHackageObject = withObject "UnresolvedPackageLocationImmutable.UPLIHackage" $ \o -> (pure) <$> (RPLIHackage
-        <$> o .: "hackage"
-        <*> o .:? "pantry-tree")
-
-optionalSubdirs' :: Object -> Parser OptionalSubdirs
-optionalSubdirs' o =
-    case HM.lookup "subdirs" o of         -- if subdirs exists, it needs to be valid
-      Just v' -> do
-        subdirs <- parseJSON v'
-        case NE.nonEmpty subdirs of
-          Nothing -> fail "Invalid empty subdirs"
-          Just x -> pure $ OSSubdirs x
-      Nothing -> OSPackageMetadata
-                 <$> o .:? "subdir" .!= T.empty
-                 <*> (RawPackageMetadata
-                      <$> (fmap unCabalString <$> (o .:? "name"))
-                      <*> (fmap unCabalString <$> (o .:? "version"))
-                      <*> o .:? "pantry-tree"
-                      <*> o .:? "cabal-file")
-
-parseRPLRepo :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseRPLRepo = withObject "UnresolvedPackageLocationImmutable.UPLIRepo" $ \o -> do
-                 (repoType, repoUrl) <-
-                     ((RepoGit, ) <$> o .: "git") <|>
-                     ((RepoHg, ) <$> o .: "hg")
-                 repoCommit <- o .: "commit"
-                 os <- optionalSubdirs' o
-                 pure $ pure $ NE.head $ NE.map (\(repoSubdir, pm) -> RPLIRepo Repo {..} pm) (osToRpms os)
-
-parseArchiveRPLObject :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseArchiveRPLObject = withObject "UnresolvedPackageLocationImmutable.UPLIArchive" $ \o -> do
-  Unresolved mkArchiveLocation <- unWarningParser $ parseArchiveLocationObject o
-  raHash <- o .:? "sha256"
-  raSize <- o .:? "size"
-  os <- optionalSubdirs' o
-  pure $ Unresolved $ \mdir -> do
-    raLocation <- mkArchiveLocation mdir
-    pure $ NE.head $ NE.map (\(raSubdir, pm) -> RPLIArchive RawArchive {..} pm) (osToRpms os)
-
-parseGithubRPLObject :: Value -> Parser (Unresolved RawPackageLocationImmutable)
-parseGithubRPLObject = withObject "PLArchive:github" $ \o -> do
-  GitHubRepo ghRepo <- o .: "github"
-  commit <- o .: "commit"
-  let raLocation = ALUrl $ T.concat
-        [ "https://github.com/"
-        , ghRepo
-        , "/archive/"
-        , commit
-        , ".tar.gz"
-        ]
-  raHash <- o .:? "sha256"
-  raSize <- o .:? "size"
-  os <- optionalSubdirs' o
-  pure $ pure $ NE.head $ NE.map (\(raSubdir, pm) -> RPLIArchive RawArchive {..} pm) (osToRpms os)
 
 instance FromJSON (WithJSONWarnings (Unresolved (NonEmpty RawPackageLocationImmutable))) where
   parseJSON v
@@ -2051,59 +1829,6 @@ instance Display SnapshotLocation where
   display (SLUrl url blob) = display url <> " (" <> display blob <> ")"
   display (SLFilePath resolved) = display (resolvedRelative resolved)
 
--- use this
-parseSL :: Value -> Parser (Unresolved SnapshotLocation)
-parseSL v = txtParser v <|> parseSLObject v
-  where
-    txt :: Text -> Maybe (Unresolved SnapshotLocation)
-    txt t = either (const Nothing) (Just . pure . SLCompiler) (parseWantedCompiler t)
-    txtParser =
-        withText
-            ("UnresolvedSnapshotLocation (Text)")
-            (\t ->
-                 pure $ fromMaybe (parseSnapshotLocationPath t) (txt t))
-
-
-
-parseSLObject :: Value -> Parser (Unresolved SnapshotLocation)
-parseSLObject = withObject "UnresolvedSnapshotLocation (Object)" $ \o ->
-                 ((pure . SLCompiler) <$> o .: "compiler") <|>
-                 ((\x y -> pure $ SLUrl x y) <$> o .: "url" <*> parseJSON (Object o)) <|>
-                 (parseSnapshotLocationPath <$> o .: "filepath")
-
--- use this
-parseRSL :: Value -> Parser (Unresolved RawSnapshotLocation)
-parseRSL v = txtParser v <|> parseRSLObject v
-    where
-      txtParser = withText "UnresolvedSnapshotLocation (Text)" (pure . parseRawSnapshotLocation)
-
-parseRSLObject :: Value -> Parser (Unresolved RawSnapshotLocation)
-parseRSLObject = withObject "UnresolvedRawSnapshotLocation (Object)" $ \o ->
-                   ((pure . RSLCompiler) <$> o .: "compiler") <|>
-                   ((\x y -> pure $ RSLUrl x y) <$> o .: "url" <*> parseBlobKey o) <|>
-                   (parseRawSnapshotLocationPath <$> o .: "filepath")
-
-
-parseBlobKey :: Object -> Parser (Maybe BlobKey)
-parseBlobKey o = do
-  msha <- o .:? "sha256"
-  msize <- o .:? "size"
-  case (msha, msize) of
-    (Nothing, Nothing) -> pure Nothing
-    (Just sha, Just size') -> pure $ Just $ BlobKey sha size'
-    (Just _sha, Nothing) -> fail "You must also specify the file size"
-    (Nothing, Just _) -> fail "You must also specify the file's SHA256"
-
-parseSnapshotLocation :: Value -> Parser (Unresolved SnapshotLocation)
-parseSnapshotLocation =
-    withObject
-        "UnresolvedSnapshotLocation"
-        (\o -> do
-             url <- o .: "url"
-             bkey <- parseJSON (Object o)
-             pure $ pure $ SLUrl url bkey)
-
-
 -- | Parse a 'Text' into an 'Unresolved' 'RawSnapshotLocation'.
 --
 -- @since 0.1.0.0
@@ -2144,15 +1869,6 @@ parseRawSnapshotLocationPath t =
     Just dir -> do
       abs' <- resolveFile dir (T.unpack t) `catchAny` \_ -> throwIO (InvalidSnapshotLocation dir t)
       pure $ RSLFilePath $ ResolvedPath (RelFilePath t) abs'
-
-parseSnapshotLocationPath :: Text -> Unresolved SnapshotLocation
-parseSnapshotLocationPath t =
-  Unresolved $ \mdir ->
-  case mdir of
-    Nothing -> throwIO $ InvalidFilePathSnapshot t
-    Just dir -> do
-      abs' <- resolveFile dir (T.unpack t) `catchAny` \_ -> throwIO (InvalidSnapshotLocation dir t)
-      pure $ SLFilePath $ ResolvedPath (RelFilePath t) abs'
 
 githubSnapshotLocation :: Text -> Text -> Text -> RawSnapshotLocation
 githubSnapshotLocation user repo path =
